Add filter

+++ app/src/main/java/com/example/lenovo/reader/fragments/article/interactors/GetArticleInteractor.kt
@@ -5,5 +5,5 @@ import io.reactivex.Single
 
 interface GetArticleInteractor {
 
-  fun execute(id: Int): Single<Article>
+  fun execute(id: Long): Single<Article>
 }
