+++ app/src/main/java/com/example/lenovo/reader/fragments/article/interactors/GetArticleInteractorImpl.kt
@@ -3,6 +3,7 @@ package com.example.lenovo.reader.fragments.article.interactors
 import com.example.model.datastores.ArticlesDataStore
 import com.example.model.models.Article
 import io.reactivex.Single
+import timber.log.Timber
 import javax.inject.Inject
 
 class GetArticleInteractorImpl @Inject constructor() : GetArticleInteractor {
@@ -10,7 +11,7 @@ class GetArticleInteractorImpl @Inject constructor() : GetArticleInteractor {
   @Inject
   lateinit var articlesDataStore: ArticlesDataStore
 
-  override fun execute(id: Int): Single<Article> {
+  override fun execute(id: Long): Single<Article> {
     return articlesDataStore.getArticle(id)
   }
